ai-project/requirements.txt
Flask==3.1.0
openai==1.68.2
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.0.1
